Chg: refacotring code.
