+++ cmd/ttid.go
@@ -1,0 +1,5 @@
+package cmd
+
+func fakeMachineID(uint16) bool {
+	return true
+}
